Move auditor.warning calls to Put/Update actions
