+++ x-pack/plugin/transform/src/main/java/org/elasticsearch/xpack/transform/action/TransportUpdateTransformAction.java
@@ -9,6 +9,7 @@ package org.elasticsearch.xpack.transform.action;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.elasticsearch.ElasticsearchStatusException;
 import org.elasticsearch.Version;
 import org.elasticsearch.action.ActionListener;
@@ -328,7 +329,12 @@ public class TransportUpdateTransformAction extends TransportTasksAction<Transfo
 
         // <3> Return to the listener
         ActionListener<Boolean> putTransformConfigurationListener = ActionListener.wrap(putTransformConfigurationResult -> {
-            auditor.info(config.getId(), "updated transform.");
+            auditor.info(config.getId(), "Updated transform.");
+            List<String> warnings = TransformWarnings.getWarnings(function, config.getSource(), config.getSyncConfig());
+            for (String warning : warnings) {
+                logger.warn(new ParameterizedMessage("[{}] {}", config.getId(), warning));
+                auditor.warning(config.getId(), warning);
+            }
             transformConfigManager.deleteOldTransformConfigurations(request.getId(), ActionListener.wrap(r -> {
                 logger.trace("[{}] successfully deleted old transform configurations", request.getId());
                 listener.onResponse(new Response(config));
